+++ app/assets/js/src/typeguardian/readTypeDef.ts
@@ -38,10 +38,11 @@ export function readTypeDef(typedef: string): TypeDef {
 		throw new Error(`Couldn't read empty type definition`);
 	}
 
-	const nameMatch = lines[0].match(/^(export\s+)?(type|interface)\s+(\w+)\s*=\s*{(\s*\/\/.*)?$/);
+	const nameMatch = lines[0].match(/^(export\s+)?(type|interface)\s+(\w+)\s*=?\s*{(\s*\/\/.*)?$/);
 	if (!nameMatch) {
 		throw new Error(`Couldn't determine name of custom type`);
 	}
+	const exported = Boolean(nameMatch[1]);
 	const name = nameMatch[3];
 
 	// Ignore first and last line
@@ -74,5 +75,6 @@ export function readTypeDef(typedef: string): TypeDef {
 	return {
 		name,
 		props,
+		exported,
 	};
 }