+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-03-31
+
+### Added
+
+* Indentation options
+* Include `export` keyword if it exists in the input
+
+### Fixed
+
+* Various UI improvements
+* Fixed error when determining name of interface
+
 ## [1.0.1] - 2023-03-31
 
 ### Fixed